+++ src/nettnc.go
@@ -338,7 +338,7 @@ func nettnc_send_packet(channel C.int, pp *packet_t) {
 	var flen = ax25_get_frame_len(pp)
 
 	var frame_buff = []byte{0} // For now, set channel to 0.
-	frame_buff = append(frame_buff, C.GoBytes(unsafe.Pointer(fbuf), flen)...)
+	frame_buff = append(frame_buff, C.GoBytes(unsafe.Pointer(fbuf), C.int(flen))...)
 
 	// Next, encapsulate into KISS frame with surrounding FENDs and any escapes.
